refactor: MVC패턴 적용 2 - racing 카테고리를 만들고 domain에 해당하는 Car와 Cars를 이동
