+++ src/main/java/racingcar/racing/domain/Cars.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.racing.domain;
 
 import java.util.List;
